Update setup.md
The name of xr_proxy container changed in the fail2ban setup instructions to match new syntax using '-' instead of '_'. Additionally, an older snode operator used the new autobuild script, a new logpath has to be queried and appended to the jails config in lines 1393 and 1403. If this is not done, the fail2ban status is changed to 'failed'. and no longer runs.